+++ src/core/ICommandProvider.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "Command.hpp"
+
+#include <vector>
+
+namespace altrun {
+
+class ICommandProvider {
+public:
+    virtual ~ICommandProvider() = default;
+
+    [[nodiscard]] virtual std::vector<Command>
+    Discover() const = 0;
+};
+
+} // namespace altrun